+++ app/lib/gen/strings_nl.g.dart
@@ -1,3 +1,9 @@
+///
+/// Generated file. Do not edit.
+///
+// coverage:ignore-file
+// ignore_for_file: type=lint
+
 part of 'strings.g.dart';
 
 // Path: <root>
@@ -538,7 +544,7 @@ class _StringsProgressPageTotalNl extends _StringsProgressPageTotalEn {
 	@override late final _StringsProgressPageTotalTitleNl title = _StringsProgressPageTotalTitleNl._(_root);
 	@override String count({required Object curr, required Object n}) => 'Bestanden: ${curr} / ${n}';
 	@override String size({required Object curr, required Object n}) => 'Grootte: ${curr} / ${n}';
-	@override String speed({required Object speed}) => 'Sneelheid: ${speed}/s';
+	@override String speed({required Object speed}) => 'Snelheid: ${speed}/s';
 }
 
 // Path: dialogs.addFile