+++ app/lib/model/state/server/receive_session_state.dart
@@ -1,9 +1,7 @@
 import 'dart:async';
 
+import 'package:common/common.dart';
 import 'package:dart_mappable/dart_mappable.dart';
-import 'package:localsend_app/model/device.dart';
-import 'package:localsend_app/model/file_type.dart';
-import 'package:localsend_app/model/session_status.dart';
 import 'package:localsend_app/model/state/server/receiving_file.dart';
 
 part 'receive_session_state.mapper.dart';
@@ -21,6 +19,7 @@ class ReceiveSessionState with ReceiveSessionStateMappable {
   final int? startTime;
   final int? endTime;
   final String destinationDirectory;
+  final String cacheDirectory;
   final bool saveToGallery;
   final StreamController<Map<String, String>?>? responseHandler;
 
@@ -33,6 +32,7 @@ class ReceiveSessionState with ReceiveSessionStateMappable {
     required this.startTime,
     required this.endTime,
     required this.destinationDirectory,
+    required this.cacheDirectory,
     required this.saveToGallery,
     required this.responseHandler,
   });