+++ app/test/mocks.mocks.dart
@@ -6,20 +6,17 @@
 import 'dart:async' as _i4;
 import 'dart:ui' as _i12;
 
+import 'package:common/common.dart' as _i2;
 import 'package:flutter/material.dart' as _i8;
 import 'package:localsend_app/gen/strings.g.dart' as _i10;
-import 'package:localsend_app/model/device.dart' as _i13;
 import 'package:localsend_app/model/persistence/color_mode.dart' as _i9;
 import 'package:localsend_app/model/persistence/favorite_device.dart' as _i6;
-import 'package:localsend_app/model/persistence/receive_history_entry.dart'
-    as _i5;
-import 'package:localsend_app/model/persistence/stored_security_context.dart'
-    as _i2;
+import 'package:localsend_app/model/persistence/receive_history_entry.dart' as _i5;
 import 'package:localsend_app/model/send_mode.dart' as _i11;
 import 'package:localsend_app/provider/persistence_provider.dart' as _i3;
 import 'package:mockito/mockito.dart' as _i1;
 import 'package:mockito/src/dummies.dart' as _i7;
-import 'package:shared_preferences/shared_preferences.dart' as _i14;
+import 'package:shared_preferences/shared_preferences.dart' as _i13;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
@@ -34,8 +31,7 @@ import 'package:shared_preferences/shared_preferences.dart' as _i14;
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
 
-class _FakeStoredSecurityContext_0 extends _i1.SmartFake
-    implements _i2.StoredSecurityContext {
+class _FakeStoredSecurityContext_0 extends _i1.SmartFake implements _i2.StoredSecurityContext {
   _FakeStoredSecurityContext_0(
     Object parent,
     Invocation parentInvocation,
@@ -48,8 +44,7 @@ class _FakeStoredSecurityContext_0 extends _i1.SmartFake
 /// A class which mocks [PersistenceService].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockPersistenceService extends _i1.Mock
-    implements _i3.PersistenceService {
+class MockPersistenceService extends _i1.Mock implements _i3.PersistenceService {
   @override
   _i2.StoredSecurityContext getSecurityContext() => (super.noSuchMethod(
         Invocation.method(
@@ -73,8 +68,7 @@ class MockPersistenceService extends _i1.Mock
       ) as _i2.StoredSecurityContext);
 
   @override
-  _i4.Future<void> setSecurityContext(_i2.StoredSecurityContext? context) =>
-      (super.noSuchMethod(
+  _i4.Future<void> setSecurityContext(_i2.StoredSecurityContext? context) => (super.noSuchMethod(
         Invocation.method(
           #setSecurityContext,
           [context],
@@ -94,8 +88,7 @@ class MockPersistenceService extends _i1.Mock
       ) as List<_i5.ReceiveHistoryEntry>);
 
   @override
-  _i4.Future<void> setReceiveHistory(List<_i5.ReceiveHistoryEntry>? entries) =>
-      (super.noSuchMethod(
+  _i4.Future<void> setReceiveHistory(List<_i5.ReceiveHistoryEntry>? entries) => (super.noSuchMethod(
         Invocation.method(
           #setReceiveHistory,
           [entries],
@@ -115,8 +108,7 @@ class MockPersistenceService extends _i1.Mock
       ) as List<_i6.FavoriteDevice>);
 
   @override
-  _i4.Future<void> setFavorites(List<_i6.FavoriteDevice>? entries) =>
-      (super.noSuchMethod(
+  _i4.Future<void> setFavorites(List<_i6.FavoriteDevice>? entries) => (super.noSuchMethod(
         Invocation.method(
           #setFavorites,
           [entries],
@@ -244,6 +236,26 @@ class MockPersistenceService extends _i1.Mock
         Invocation.method(
           #setPort,
           [port],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+
+  @override
+  bool getShareViaLinkAutoAccept() => (super.noSuchMethod(
+        Invocation.method(
+          #getShareViaLinkAutoAccept,
+          [],
+        ),
+        returnValue: false,
+        returnValueForMissingStub: false,
+      ) as bool);
+
+  @override
+  _i4.Future<void> setShareViaLinkAutoAccept(bool? shareViaLinkAutoAccept) => (super.noSuchMethod(
+        Invocation.method(
+          #setShareViaLinkAutoAccept,
+          [shareViaLinkAutoAccept],
         ),
         returnValue: _i4.Future<void>.value(),
         returnValueForMissingStub: _i4.Future<void>.value(),
@@ -382,8 +394,7 @@ class MockPersistenceService extends _i1.Mock
       ) as bool);
 
   @override
-  _i4.Future<void> setMinimizeToTray(bool? minimizeToTray) =>
-      (super.noSuchMethod(
+  _i4.Future<void> setMinimizeToTray(bool? minimizeToTray) => (super.noSuchMethod(
         Invocation.method(
           #setMinimizeToTray,
           [minimizeToTray],
@@ -403,8 +414,7 @@ class MockPersistenceService extends _i1.Mock
       ) as bool);
 
   @override
-  _i4.Future<void> setLaunchAtStartup(bool? launchAtStartup) =>
-      (super.noSuchMethod(
+  _i4.Future<void> setLaunchAtStartup(bool? launchAtStartup) => (super.noSuchMethod(
         Invocation.method(
           #setLaunchAtStartup,
           [launchAtStartup],
@@ -424,8 +434,7 @@ class MockPersistenceService extends _i1.Mock
       ) as bool);
 
   @override
-  _i4.Future<void> setAutoStartLaunchMinimized(bool? launchMinimized) =>
-      (super.noSuchMethod(
+  _i4.Future<void> setAutoStartLaunchMinimized(bool? launchMinimized) => (super.noSuchMethod(
         Invocation.method(
           #setAutoStartLaunchMinimized,
           [launchMinimized],
@@ -515,8 +524,7 @@ class MockPersistenceService extends _i1.Mock
       ) as _i4.Future<void>);
 
   @override
-  Map<String, _i12.OffsetBase?> getWindowLastDimensions() =>
-      (super.noSuchMethod(
+  Map<String, _i12.OffsetBase?> getWindowLastDimensions() => (super.noSuchMethod(
         Invocation.method(
           #getWindowLastDimensions,
           [],
@@ -526,8 +534,7 @@ class MockPersistenceService extends _i1.Mock
       ) as Map<String, _i12.OffsetBase?>);
 
   @override
-  _i4.Future<void> setSaveWindowPlacement(bool? savePlacement) =>
-      (super.noSuchMethod(
+  _i4.Future<void> setSaveWindowPlacement(bool? savePlacement) => (super.noSuchMethod(
         Invocation.method(
           #setSaveWindowPlacement,
           [savePlacement],
@@ -547,8 +554,7 @@ class MockPersistenceService extends _i1.Mock
       ) as bool);
 
   @override
-  _i4.Future<void> setEnableAnimations(bool? enableAnimations) =>
-      (super.noSuchMethod(
+  _i4.Future<void> setEnableAnimations(bool? enableAnimations) => (super.noSuchMethod(
         Invocation.method(
           #setEnableAnimations,
           [enableAnimations],
@@ -568,8 +574,7 @@ class MockPersistenceService extends _i1.Mock
       ) as bool);
 
   @override
-  _i4.Future<void> setDeviceType(_i13.DeviceType? deviceType) =>
-      (super.noSuchMethod(
+  _i4.Future<void> setDeviceType(_i2.DeviceType? deviceType) => (super.noSuchMethod(
         Invocation.method(
           #setDeviceType,
           [deviceType],
@@ -602,7 +607,7 @@ class MockPersistenceService extends _i1.Mock
 /// A class which mocks [SharedPreferences].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockSharedPreferences extends _i1.Mock implements _i14.SharedPreferences {
+class MockSharedPreferences extends _i1.Mock implements _i13.SharedPreferences {
   @override
   Set<String> getKeys() => (super.noSuchMethod(
         Invocation.method(