+++ app/assets/CHANGELOG.md
@@ -1,6 +1,15 @@
-## 1.14.0 (unreleased)
+## 1.15.0 (unreleased)
 
+- fix(windows): correctly remove tray icon when closing the app (@zpp0196)
+- i18n: add Serbian (@nebojsatomic)
+
+## 1.14.0 (2024-02-26)
+
+- feat: add option to automatically accept requests when sharing via link (@MisterChangRay, @Tienisto)
+- feat: use fix button width for all buttons in the selection row (only noticeable in Russian) (@Tienisto)
 - fix: picking many files should not freeze the UI (@Tienisto)
+- fix: do not create a new session for the same IP when sharing via link (@MisterChangRay)
+- fix(android): save files to SD card on Android 10 or older (@Tienisto)
 - i18n: add Danish (@Limfjorden)
 
 ## 1.13.1 (2023-12-08)